modify sample vanDerPol FMU parameters to support control
The original vanDerPol FMU ran without being controlled by external input. For training a Bonsai brain, we need it to support actions and states.

Instead of overriding the FMU's variable definitions using a config file in the FMU connector, this change switches the variable causality in the modelDescription.xml of the FMU itself.

x0 is changed to input. der(x0), x1, and derx1 are changed to output.

With this change, vanDerPol_conf.yaml is no longer necessary and can be removed. The default variable types read from the FMU are sufficient.